# alef_delta_sacco_backend/openapi.yaml
openapi: 3.0.3
info:
  title: ALEF-DELTA SACCO API
  version: 1.0.0
  description: |
    Complete backend API for ALEF-DELTA SACCO Management System.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token.
    
    ## Roles
    - **ADMIN**: Full system access
    - **TELLER**: Cash operations, member creation
    - **CREDIT_OFFICER**: Loan processing, collateral/guarantor verification
    - **MANAGER**: Approvals, member activation
    - **AUDITOR**: Read-only access
    - **MEMBER**: Client self-service
servers:
  - url: http://localhost:4000/api
    description: Local development server
  - url: https://api.alef-delta.local/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Member:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
        membership_no:
          type: string
        first_name:
          type: string
        middle_name:
          type: string
        last_name:
          type: string
        phone_primary:
          type: string
        email:
          type: string
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
        marital_status:
          type: string
          enum: [SINGLE, MARRIED, DIVORCED, WIDOWED]
        member_type:
          type: string
          enum: [INDIVIDUAL, GOV_EMP, NGO, SME]
        monthly_income:
          type: number
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, CLOSED]
    Account:
      type: object
      properties:
        account_id:
          type: string
        member_id:
          type: string
        product_code:
          type: string
        balance:
          type: number
        lien_amount:
          type: number
        available_balance:
          type: number
        currency:
          type: string
        status:
          type: string
        version:
          type: integer
...
paths:
...
  /members:
    get:
      summary: List members (Staff only)
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, SUSPENDED, CLOSED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Members list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Member'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                    
    post:
      summary: Create member (Teller/Manager/Admin)
      tags: [Members]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [first_name, last_name, phone_primary, gender, marital_status, member_type, password]
      responses:
        '201':
          description: Member created
  /members/{memberId}:
    get:
      summary: Get member details
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
                
    put:
      summary: Update member (Manager/Admin)
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Member'
      responses:
        '200':
          description: Member updated
          
    delete:
      summary: Delete member (Admin only)
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member deleted
  /members/{memberId}/activate:
    post:
      summary: Activate member (Manager/Admin)
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member activated
  /members/{memberId}/suspend:
    post:
      summary: Suspend member (Manager/Admin)
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Member suspended
  /members/{memberId}/upload:
    post:
      summary: Upload member documents
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                profile_photo:
                  type: string
                  format: binary
                id_card:
                  type: string
                  format: binary
      responses:
        '200':
          description: Documents uploaded
  /members/{memberId}/beneficiaries:
    post:
      summary: Add beneficiary
      tags: [Members]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name, relationship, phone]
              properties:
                full_name:
                  type: string
                relationship:
                  type: string
                phone:
                  type: string
                id_front:
                  type: string
                  format: binary
                id_back:
                  type: string
                  format: binary
      responses:
        '201':
          description: Beneficiary added
          
  # ==================== ACCOUNTS ====================
...
  /accounts/member/{memberId}:
    get:
      summary: Get accounts for a member
      tags: [Accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
...